# openapi.yaml
openapi: 3.0.1
info:
  title: AskYourCode
  description: A plugin that provides access to the source code and documentation of a software project provided by the user
  version: 'v1'
servers:
  - url: https://plugin.askyourcode.ai
paths:
  /project/{username}:
    post:
      operationId: downloadProject
      summary: Downloads the source code archive of the project, extracts the archive, splits the source code and documentation into fragments, stores the fragments into a vector database
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Name of the logged in user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/downloadRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/downloadResponse'
  /project/{username}/{project_id}:
    delete:
      operationId: deleteProject
      summary: Deletes a project from the vector database
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Name of the logged in user
        - in: path
          name: project_id
          schema:
            type: string
          required: true
          description: GUID of the project
      responses:
        "200":
          description: OK
  /project/{username}/{project_id}/search:
    get:
      operationId: search
      summary: Search for code or documentation fragments
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Name of the logged in user
        - in: path
          name: project_id
          schema:
            type: string
          required: true
          description: GUID of the project
        - in: query
          name: text
          schema:
            type: string
          required: true
          description: Query string for the vector database, start with .extension to narrow down to one file type
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 3
          required: false
          description: Maximum number of results to return
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/searchResponse'

components:
  schemas:
    downloadRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          description: The URL of the source code archive to download. Must point to a zip or tar.gz file which can be downloaded directly without authentication.
          required: true
    downloadResponse:
      type: object
      required:
        - project_id
      properties:
        project_id:
          type: string
          description: GUID of the project
          required: true
        error:
          type: string
          description: Error message (missing if the request was successful)
          required: false
    searchResponse:
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            type: object
            required:
              - score
              - text
              - meta
            properties:
              score:
                type: number
                description: Score, higher value means more relevant
              text:
                type: string
                description: Text fragment
              meta:
                type: object
                required:
                  - path
                properties:
                  path:
                    type: string
                    description: Relative path inside the project
                  kind:
                    type: string
                    description: Language construct (module, class, function, method, interface, etc.)
                  name:
                    type: string
                    description: Name of the definition or declaration
                  index:
                    type: integer
                    description: Index to order the fragments inside the same definition
                description: Fragment meta-information
          description: Search results in order of decreasing score
          required: true